Testing our TwoDShape Abastract class and subclassses in TwoDShapeMain with main method --> note an advantange of using abstract class as we easily change Volume to perimeter in the shapeInfo method

+++ src/abstractClassRealExample/Circle.java
@@ -2,11 +2,11 @@ package abstractClassRealExample;
 
 public class Circle extends TwoDShapes{
 	private static double PI = 3.142;
-	private int radius;
+	private double  radius;
 	
 	
 
-	public Circle(String name, int radius) {
+	public Circle(String name, double radius) {
 		super(name);
 		this.radius = radius;
 	}
